Session

+++ server/controller/Cerrands.js
@@ -1,6 +1,5 @@
-const Errands = require('../models');
-const { Op } = require('sequelize');
-const { checkout } = require('../routes/errands');
+const Errands = require("../models");
+const { Op } = require("sequelize");
 // Access => Errands.User_info or Helper_board etc
 
 // ======= User sign =======
@@ -14,6 +13,12 @@ exports.userLogin = async (req, res) => {
       },
     });
     if (result != null) {
+      req.session.user_name = result.dataValues.user_name;
+      req.session.user_type = result.dataValues.user_type;
+      console.log(result.dataValues.user_name);
+      console.log(req.session.user_name);
+      console.log(req.session.user_type);
+      console.log(req.session);
       res.send(true);
     } else {
       res.send(false);
@@ -46,10 +51,10 @@ exports.checkUserName = async (req, res) => {
     const result = await Errands.User_info.findOne({
       where: { user_name: { [Op.eq]: req.body.user_name } },
     });
-    if (result.dataValues.user_name !== null) {
-      res.send(false);
-    } else {
+    if (!result) {
       res.send(true);
+    } else {
+      res.send(false);
     }
   } catch (err) {
     res.send(err);
@@ -60,7 +65,7 @@ exports.checkUserName = async (req, res) => {
 exports.userRegister = async (req, res) => {
   try {
     const result = await Errands.User_info.findOne({
-      where: { user_id: req.body.user_id },
+      where: { user_id: { [Op.eq]: req.body.user_id } },
     });
     console.log(result);
     if (!result) {
@@ -76,12 +81,76 @@ exports.userRegister = async (req, res) => {
   }
 };
 
+// 로그아웃
+exports.userLogout = async (req, res) => {
+  try {
+    await req.session.destroy((err) => {
+      if (err) {
+        throw err;
+      }
+      req.session;
+      res.send(true);
+    });
+  } catch (err) {
+    res.send(err);
+  }
+};
+
+// 메인페이지 상위 5명 보여주기
+exports.read_few_user = async (req, res) => {
+  try {
+    const result = await Errands.User_info.findAll({
+      order: [["user_like", "desc"]],
+      limit: 5,
+    });
+    res.send(result);
+  } catch (err) {
+    res.send(err);
+  }
+};
+// 전체 다 보여주기
+exports.read_user = async (req, res) => {
+  try {
+    const result = await Errands.User_info.findAll({
+      order: [["user_like", "desc"]],
+    });
+    res.send(result);
+  } catch (err) {
+    res.send(err);
+  }
+};
+
+// detail
+exports.read_detail_user = async (req, res) => {
+  try {
+    const result = await Errands.User_info.findOne({
+      where: { id: { [Op.eq]: req.params.user } },
+    });
+    res.send(result);
+  } catch (err) {
+    res.send(err);
+  }
+};
+
+// 추천수
+exports.userLike = async (req, res) => {
+  try {
+    const result = await Errands.User_info.increment(
+      { user_like: 1 },
+      { where: { id: { [Op.eq]: req.params.user } } }
+    );
+    res.send(result);
+  } catch (err) {
+    res.send(err);
+  }
+};
+
 // ======= Wanter_board =======
 // 매인페이지에 5개 보여주기 deadline순 5개
 exports.read_few_wanter_board = async (req, res) => {
   try {
     const result = await Errands.Wanter_board.findAll({
-      order: [['wanter_board_dead_line', 'desc']],
+      order: [["wanter_board_dead_line", "desc"]],
       limit: 5,
     });
     res.send(result);
@@ -94,7 +163,7 @@ exports.read_few_wanter_board = async (req, res) => {
 exports.read_wanter_board = async (req, res) => {
   try {
     const result = await Errands.Wanter_board.findAll({
-      order: [['wanter_board_date', 'asc']],
+      order: [["wanter_board_date", "asc"]],
     });
     res.send(result);
 
@@ -124,7 +193,7 @@ exports.create_wanter_board = async (req, res) => {
       wanter_board_title: req.body.wanter_board_title,
       wanter_board_content: req.body.wanter_board_content,
       wanter_board_place: req.body.wanter_board_place,
-      wanter_board_dead_line: req.body.wanter_board_dead_line.value,
+      wanter_board_dead_line: req.body.wanter_board_dead_line,
       wanter_board_done: false,
     });
     res.send(result);
@@ -135,23 +204,29 @@ exports.create_wanter_board = async (req, res) => {
 
 // 게시물 수정
 exports.update_wanter_board = async (req, res) => {
-  console.log('**********************8');
-  console.log(req.params.boardId);
-  console.log(req.params.wanter_board_id);
   try {
-    const [result] = await Errands.Wanter_board.update(
-      {
-        wanter_board_title: req.body.wanter_board_title,
-        wanter_board_content: req.body.wanter_board_place,
-        wanter_board_place: req.body.wanter_board_place,
-        wanter_board_done: req.body.done,
-      },
-      { where: { wanter_board_id: { [Op.eq]: req.params.boardId } } }
-    );
-    if (result === 0) {
-      return res.send(false);
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_name"],
+      where: { user_name: { [Op.eq]: req.session.user_name } },
+    });
+    console.log(auth);
+    if (auth.dataValues.user_name == req.session.user_name) {
+      const [result] = Errands.Wanter_board.update(
+        {
+          wanter_board_title: req.body.wanter_board_title,
+          wanter_board_content: req.body.wanter_board_place,
+          wanter_board_place: req.body.wanter_board_place,
+          wanter_board_done: req.body.done,
+        },
+        { where: { wanter_board_id: { [Op.eq]: req.params.boardId } } }
+      );
+      if (result === 0) {
+        return res.send(false);
+      }
+      res.send(true);
+    } else {
+      res.send("작성자만 수정할 수 있습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
@@ -160,13 +235,21 @@ exports.update_wanter_board = async (req, res) => {
 // 게시물 삭제
 exports.delete_wanter_board = async (req, res) => {
   try {
-    const result = await Errands.Wanter_board.destroy({
-      where: { wanter_board_id: { [Op.eq]: req.params.boardId } },
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_name"],
+      where: { user_name: { [Op.eq]: req.session.user_name } },
     });
-    if (!result) {
-      return res.send(false);
+    if (auth.dataValues.user_name == req.session.user_name) {
+      const result = Errands.Wanter_board.destroy({
+        where: { wanter_board_id: { [Op.eq]: req.params.boardId } },
+      });
+      if (!result) {
+        return res.send(false);
+      }
+      res.send(true);
+    } else {
+      res.send("작성자만 삭제할 수 있습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
@@ -176,7 +259,9 @@ exports.delete_wanter_board = async (req, res) => {
 // 댓글 보여주기
 exports.read_wanter_comment = async (req, res) => {
   try {
-    const result = await Errands.Wanter_comment.findAll();
+    const result = await Errands.Wanter_comment.findAll({
+      where: { wanter_comment_board_id: { [Op.eq]: req.params.boardId } },
+    });
     res.send(result);
   } catch (err) {
     res.send(err);
@@ -188,7 +273,7 @@ exports.create_wanter_comment = async (req, res) => {
   try {
     const result = await Errands.Wanter_comment.create({
       wanter_comment_board_id: req.params.boardId,
-      wanter_comment_writer: req.body.user_name,
+      wanter_comment_writer: req.session.user_name,
       wanter_comment_content: req.body.wanter_comment_content,
     });
     res.send(result);
@@ -200,21 +285,29 @@ exports.create_wanter_comment = async (req, res) => {
 // 댓글 수정
 exports.update_wanter_comment = async (req, res) => {
   try {
-    const [result] = await Errands.Wanter_comment.update(
-      {
-        wanter_comment_content: req.body.wanter_comment_content,
-      },
-      {
-        where: {
-          wanter_comment_id: { [Op.eq]: req.params.commentId },
-          wanter_comment_board_id: { [Op.eq]: req.params.boardId },
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_name"],
+      where: { user_name: { [Op.eq]: req.session.user_name } },
+    });
+    if (auth.dataValues.user_name == req.session.user_name) {
+      const [result] = Errands.Wanter_comment.update(
+        {
+          wanter_comment_content: req.body.wanter_comment_content,
         },
+        {
+          where: {
+            wanter_comment_id: { [Op.eq]: req.params.commentId },
+            wanter_comment_board_id: { [Op.eq]: req.params.boardId },
+          },
+        }
+      );
+      if (result === 0) {
+        return res.send(false);
       }
-    );
-    if (result === 0) {
-      return res.send(false);
+      res.send(true);
+    } else {
+      res.send("작성자만 삭제할 수 있습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
@@ -223,16 +316,24 @@ exports.update_wanter_comment = async (req, res) => {
 // 댓글 삭제
 exports.delete_wanter_comment = async (req, res) => {
   try {
-    const result = await Errands.Wanter_comment.destroy({
-      where: {
-        wanter_comment_id: { [Op.eq]: req.params.commentId },
-        wanter_comment_board_id: { [Op.eq]: req.params.boardId },
-      },
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_name"],
+      where: { user_name: { [Op.eq]: req.session.user_name } },
     });
-    if (!result) {
-      return res.send(false);
+    if (auth.dataValues.user_name == req.session.user_name) {
+      const result = Errands.Wanter_comment.destroy({
+        where: {
+          wanter_comment_id: { [Op.eq]: req.params.commentId },
+          wanter_comment_board_id: { [Op.eq]: req.params.boardId },
+        },
+      });
+      if (!result) {
+        return res.send(false);
+      }
+      res.send(true);
+    } else {
+      res.send("작성자만 댓글 삭제할 수 있습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
@@ -243,7 +344,7 @@ exports.delete_wanter_comment = async (req, res) => {
 exports.read_few_helper_board = async (req, res) => {
   try {
     const result = await Errands.Helper_board.findAll({
-      order: [['helper_board_date', 'asc']],
+      order: [["helper_board_date", "asc"]],
       limit: 3,
     });
     res.send(result);
@@ -256,7 +357,7 @@ exports.read_few_helper_board = async (req, res) => {
 exports.read_helper_board = async (req, res) => {
   try {
     const result = await Errands.Helper_board.findAll({
-      helper_board_writer: req.body.helper_board_writer,
+      helper_board_writer: req.session.user_name,
       helper_board_title: req.body.helper_board_title,
       helper_board_content: req.body.helper_board_content,
       helper_board_place: req.body.helper_board_place,
@@ -284,7 +385,7 @@ exports.read_one_helper_board = async (req, res) => {
 exports.create_helper_board = async (req, res) => {
   try {
     const result = await Errands.Helper_board.create({
-      helper_board_writer: req.body.helper_board_writer,
+      helper_board_writer: req.session.user_name,
       helper_board_title: req.body.helper_board_title,
       helper_board_content: req.body.helper_board_content,
       helper_board_place: req.body.helper_board_place,
@@ -299,21 +400,29 @@ exports.create_helper_board = async (req, res) => {
 // 게시물 수정
 exports.update_helper_board = async (req, res) => {
   try {
-    const [result] = await Errands.Helper_board.update(
-      {
-        helper_board_title: req.body.helper_board_title,
-        helper_board_content: req.body.helper_board_content,
-        helper_board_place: req.body.helper_board_place,
-        helper_board_done: req.body.done,
-      },
-      {
-        where: { helper_board_id: { [Op.eq]: req.params.boardId } },
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_name"],
+      where: { user_name: { [Op.eq]: req.session.user_name } },
+    });
+    if (auth.dataValues.user_name == req.session.user_name) {
+      const [result] = Errands.Helper_board.update(
+        {
+          helper_board_title: req.body.helper_board_title,
+          helper_board_content: req.body.helper_board_content,
+          helper_board_place: req.body.helper_board_place,
+          helper_board_done: req.body.done,
+        },
+        {
+          where: { helper_board_id: { [Op.eq]: req.params.boardId } },
+        }
+      );
+      if (result === 0) {
+        return res.send(false);
       }
-    );
-    if (result === 0) {
-      return res.send(false);
+      res.send(true);
+    } else {
+      res.send("작성자만 수정할 수 있습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
@@ -322,13 +431,21 @@ exports.update_helper_board = async (req, res) => {
 // 게시물 삭제
 exports.delete_helper_board = async (req, res) => {
   try {
-    const result = await Errands.Helper_board.destroy({
-      where: { helper_board_id: { [Op.eq]: req.params.boardId } },
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_name"],
+      where: { user_name: { [Op.eq]: req.session.user_name } },
     });
-    if (!result) {
-      return res.send(false);
+    if (auth.dataValues.user_name == req.session.user_name) {
+      const result = Errands.Helper_board.destroy({
+        where: { helper_board_id: { [Op.eq]: req.params.boardId } },
+      });
+      if (!result) {
+        return res.send(false);
+      }
+      res.send(true);
+    } else {
+      res.send("작성자만 삭제할 수 있습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
@@ -339,7 +456,7 @@ exports.delete_helper_board = async (req, res) => {
 exports.read_helper_comment = async (req, res) => {
   try {
     const result = await Errands.Helper_comment.findAll({
-      //
+      where: { wanter_comment_board_id: { [Op.eq]: req.params.boardId } },
     });
     res.send(result);
   } catch (err) {
@@ -352,7 +469,7 @@ exports.create_helper_comment = async (req, res) => {
   try {
     const result = await Errands.Helper_comment.create({
       helper_comment_board_id: req.params.boardId,
-      helper_comment_writer: req.body.helper_comment_writer,
+      helper_comment_writer: req.session.user_name,
       helper_comment_content: req.body.helper_comment_content,
     });
     res.send(result);
@@ -364,21 +481,29 @@ exports.create_helper_comment = async (req, res) => {
 // 댓글 수정
 exports.update_helper_comment = async (req, res) => {
   try {
-    const [result] = await Errands.Helper_comment.update(
-      {
-        helper_comment_content: req.body.helper_comment_content,
-      },
-      {
-        where: {
-          helper_comment_id: { [Op.eq]: req.params.commentId },
-          helper_comment_board_id: { [Op.eq]: req.params.boardId },
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_name"],
+      where: { user_name: { [Op.eq]: req.session.user_name } },
+    });
+    if (auth.dataValues.user_name == req.session.user_name) {
+      const [result] = Errands.Helper_comment.update(
+        {
+          helper_comment_content: req.body.helper_comment_content,
         },
+        {
+          where: {
+            helper_comment_id: { [Op.eq]: req.params.commentId },
+            helper_comment_board_id: { [Op.eq]: req.params.boardId },
+          },
+        }
+      );
+      if (result === 0) {
+        return res.send(false);
       }
-    );
-    if (result === 0) {
-      return res.send(false);
+      res.send(true);
+    } else {
+      res.send("작성자만 수정할 수 있습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
@@ -387,16 +512,24 @@ exports.update_helper_comment = async (req, res) => {
 // 댓글 삭제
 exports.delete_helper_comment = async (req, res) => {
   try {
-    const result = await Errands.Helper_comment.destroy({
-      where: {
-        helper_comment_id: { [Op.eq]: req.params.commentId },
-        helper_comment_board_id: { [Op.eq]: req.params.boardId },
-      },
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_name"],
+      where: { user_name: { [Op.eq]: req.session.user_name } },
     });
-    if (!result) {
-      return res.send(false);
+    if (auth.dataValues.user_name == req.session.user_name) {
+      const result = Errands.Helper_comment.destroy({
+        where: {
+          helper_comment_id: { [Op.eq]: req.params.commentId },
+          helper_comment_board_id: { [Op.eq]: req.params.boardId },
+        },
+      });
+      if (!result) {
+        return res.send(false);
+      }
+      res.send(true);
+    } else {
+      res.send("작성자만 삭제할 수 있습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
@@ -408,7 +541,7 @@ exports.delete_helper_comment = async (req, res) => {
 exports.read_few_notice = async (req, res) => {
   try {
     const result = await Errands.Notice.findAll({
-      order: [['notice_date', 'desc']],
+      order: [["notice_date", "desc"]],
       limit: 5,
     });
     res.send(result);
@@ -421,7 +554,7 @@ exports.read_few_notice = async (req, res) => {
 exports.read_notice = async (req, res) => {
   try {
     const result = await Errands.Notice.findAll({
-      order: [['wanter_board_date', 'asc']],
+      order: [["wanter_board_date", "asc"]],
     });
   } catch (err) {
     res.send(err);
@@ -443,12 +576,21 @@ exports.read_one_notice = async (req, res) => {
 // 공지 생성
 exports.create_notice = async (req, res) => {
   try {
-    const result = await Errands.notice.create({
-      notice_writer: req.body.notice_writer,
-      notice_title: req.body.notice_title,
-      notice_content: req.body.notice_content,
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_type"],
+      where: { user_type: { [Op.eq]: req.session.user_type } },
     });
-    res.send(result);
+    if (auth.dataValues.user_type == "root") {
+      const result = Errands.Notice.create({
+        notice_writer: req.session.user_name,
+        notice_title: req.body.notice_title,
+        notice_content: req.body.notice_content,
+      });
+
+      res.send(result);
+    } else {
+      res.send("생성 권한이 없습니다");
+    }
   } catch (err) {
     res.send(err);
   }
@@ -457,17 +599,25 @@ exports.create_notice = async (req, res) => {
 // 공지 수정
 exports.update_notice = async (req, res) => {
   try {
-    const [result] = await Errands.Notice.update(
-      {
-        notice_title: req.body.notice_title,
-        notice_content: req.body.notice_content,
-      },
-      { where: { notice_id: { [Op.eq]: req.params.boardId } } }
-    );
-    if (result === 0) {
-      return res.send(false);
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_type"],
+      where: { user_type: { [Op.eq]: req.session.user_type } },
+    });
+    if (auth.dataValues.user_type == "root") {
+      const [result] = Errands.Notice.update(
+        {
+          notice_title: req.body.notice_title,
+          notice_content: req.body.notice_content,
+        },
+        { where: { notice_id: { [Op.eq]: req.params.boardId } } }
+      );
+      if (result === 0) {
+        return res.send(false);
+      }
+      res.send(true);
+    } else {
+      res.send("수정 권한이 없습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
@@ -476,13 +626,21 @@ exports.update_notice = async (req, res) => {
 // 공지 삭제
 exports.delete_notice = async (req, res) => {
   try {
-    const result = await Errands.Notice.destroy({
-      where: { notice_id: { [Op.eq]: req.params.boardId } },
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_type"],
+      where: { user_type: { [Op.eq]: req.session.user_type } },
     });
-    if (!result) {
-      return res.send(false);
+    if (auth.dataValues.user_type == "root") {
+      const result = Errands.Notice.destroy({
+        where: { notice_id: { [Op.eq]: req.params.boardId } },
+      });
+      if (!result) {
+        return res.send(false);
+      }
+      res.send(true);
+    } else {
+      res.send("삭제 권한이 없습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
